removed deviceid for all queries in graphs, added jwt. added ws for today so it updated when correctbroadcast is used.

+++ server/Core.Domain/Entities/Devicelog.cs
@@ -19,5 +19,5 @@ public partial class Devicelog
 
     public decimal Pressure { get; set; }
 
-    public int Airquality { get; set; }
+    public double Airquality { get; set; }
 }
